% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta_carotene_data}
\alias{beta_carotene_data}
\title{Beta-carotene RoB Judgments (Raw)
Domain-level risk of bias judgments for 15 studies of beta-carotene and cardiovascular outcomes.}
\format{
A data frame with 27 rows and 25 variables.
}
\source{
Supplement to Shapland et al. (2024)
}
\usage{
beta_carotene_data
}
\description{
Beta-carotene RoB Judgments (Raw)
Domain-level risk of bias judgments for 15 studies of beta-carotene and cardiovascular outcomes.
}
\keyword{datasets}
