% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_bias}
\alias{dat_bias}
\title{Example risk of bias assessments}
\format{
A data frame with 20 rows and 26 variables
}
\usage{
dat_bias
}
\description{
Example risk of bias assessments
}
\keyword{datasets}
