% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_bias_values}
\alias{dat_bias_values}
\title{Example priors for the bias adjustment}
\format{
A data frame with 3 rows and 6 variables
}
\usage{
dat_bias_values
}
\description{
Example priors for the bias adjustment
}
\keyword{datasets}
