% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_CHD_bias}
\alias{dat_CHD_bias}
\title{Beta-carotene CHD data (with RoB)
Cleaned dataset of log risk ratios for CHD, merged with RoB annotations.}
\format{
A data frame with 13 rows and ~25 variables
}
\source{
Shapland et al. (2024)
}
\usage{
dat_CHD_bias
}
\description{
Beta-carotene CHD data (with RoB)
Cleaned dataset of log risk ratios for CHD, merged with RoB annotations.
}
\keyword{datasets}
