% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_CVD_bias}
\alias{dat_CVD_bias}
\title{Beta-carotene CVD data (with RoB)
Cleaned dataset of log risk ratios for CVD, merged with RoB annotations.}
\format{
A data frame with 17 rows and ~25 variables
}
\source{
Shapland et al. (2024)
}
\usage{
dat_CVD_bias
}
\description{
Beta-carotene CVD data (with RoB)
Cleaned dataset of log risk ratios for CVD, merged with RoB annotations.
}
\keyword{datasets}
