% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_ind}
\alias{dat_ind}
\title{Example indirectness assessments}
\format{
A data frame with 20 rows and 14 variables
}
\usage{
dat_ind
}
\description{
Example indirectness assessments
}
\keyword{datasets}
