% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_ind_values}
\alias{dat_ind_values}
\title{Example priors for indirectness adjustment}
\format{
A data frame with 3 rows and 6 variables
}
\usage{
dat_ind_values
}
\description{
Example priors for indirectness adjustment
}
\keyword{datasets}
