% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_triangles.R
\name{plot_triangles}
\alias{plot_triangles}
\title{Visualise the points and edges from triangulate}
\usage{
plot_triangles(plot_object)
}
\arguments{
\item{plot_object}{output from the `triangulate()` function}
}
\value{
A ggplot2 object
}
\description{
This function plots the points and edges from the
triangulate function to visualise the triangles
generated by connecting the five points
}
\examples{
points <- matrix(c(0,0,1,1,2,0,3,1,4,0), nrow=5, byrow=T)
results <- triangulate(points)
plot_triangles(results)
}
