% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data1}
\alias{sample_data1}
\title{Sample Dataset 1: Uniformly Distributed Points}
\format{
A numeric matrix with 5 rows and 2 columns:
\describe{
  \item{x}{X-coordinates}
  \item{y}{Y-coordinates}
}
}
\usage{
sample_data1
}
\description{
A 5x2 matrix of five points with x and y coordinates, uniformly distributed
from 0 to 10
}
\keyword{datasets}
