% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data2}
\alias{sample_data2}
\title{Sample Dataset 2: Much wider Points with some Clustering}
\format{
A numeric matrix with 5 rows and 2 columns:
\describe{
  \item{x}{X-coordinate}
  \item{y}{Y-coordinate}
}
}
\usage{
sample_data2
}
\description{
A 5x2 matrix of five points with x and y coordinates, spread across a range
of -10 to 10, with some clustering around the origin.
}
\keyword{datasets}
