% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_direction.R
\name{tri_absolute_direction}
\alias{tri_absolute_direction}
\title{Convert to absolute direction of bias or indirectness}
\usage{
tri_absolute_direction(dat)
}
\arguments{
\item{dat}{Data in long format}
}
\description{
Converts qualitative bias direction labels (e.g. "Favours comparator")
into absolute directions ("left" or "right") based on effect size and bias type.
}
