% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_direction.R
\name{tri_absolute_direction_invert}
\alias{tri_absolute_direction_invert}
\title{Convert absolute direction back to qualitative tool labels}
\usage{
tri_absolute_direction_invert(dat)
}
\arguments{
\item{dat}{Data in long format}
}
\description{
Reverses the transformation done by \code{tri_absolute_direction()}, converting
absolute directions ("left"/"right") back to labels like "Favours comparator".
}
