% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_direction.R
\name{tri_absolute_direction_quick}
\alias{tri_absolute_direction_quick}
\title{Title}
\usage{
tri_absolute_direction_quick(dat, invert = FALSE)
}
\arguments{
\item{dat}{Triangulation dataset}

\item{invert}{Direction indicator}
}
\description{
A helper function to quickly convert from relative to absolute
directions of bias, or vice versa
}
