% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_values.R
\name{tri_append_bias}
\alias{tri_append_bias}
\title{Title}
\usage{
tri_append_bias(dat, values, common = T)
}
\arguments{
\item{dat}{Bias dataframe in long format}

\item{values}{Distributions of bias for different risk of bias levels}

\item{common}{Should a single set of distributions be used across all domains
(default is TRUE)?}
}
\description{
Title
}
