% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_values.R
\name{tri_append_indirect}
\alias{tri_append_indirect}
\title{Title}
\usage{
tri_append_indirect(dat, values, common = T)
}
\arguments{
\item{dat}{Indirectness dataframe in long format}

\item{values}{Distributions of bias for different risk of indirectness levels}

\item{common}{Should a single set of distributions be used across all domains
(default is TRUE)?}
}
\description{
Title
}
