% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducible_plot.R
\name{tri_plot_bias_direction01}
\alias{tri_plot_bias_direction01}
\title{Plot bias-adjusted and unadjusted estimates with RoB, adaptable version of tri_plot_bias_direction}
\usage{
tri_plot_bias_direction01(
  dat,
  dat_adj = NULL,
  title = NULL,
  grouping = "type",
  rma_method = "REML",
  ...
)
}
\arguments{
\item{dat}{A dataframe output from \code{tri_calculate_adjusted_estimates()}, with \code{yi}, \code{vi}, \code{yi_adj}, \code{vi_adj}, and domain columns.}

\item{dat_adj}{A data frame with adjusted estimates and variances, if different from the main input.}

\item{title}{Optional plot title.}

\item{grouping}{Column by which to stratify subgroups (default = \code{"type"}).}

\item{rma_method}{Meta-analysis method passed to metafor::rma (e.g., "REML", "FE").}

\item{...}{Additional arguments passed to \code{rob_direction()}.}
}
\value{
A forest plot is drawn (base graphics).
}
\description{
Creates a forest plot of both raw and adjusted effect estimates, alongside domain-level
risk of bias annotations for each study.
}
