% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{tri_prep_data}
\alias{tri_prep_data}
\title{Title}
\usage{
tri_prep_data(dat_bias, dat_ind)
}
\arguments{
\item{dat_bias}{Cleaned bias dataset}

\item{dat_ind}{Cleaned indirectness dataset}
}
\description{
Title
}
