% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_direction.R
\name{tri_swap_effect_direction}
\alias{tri_swap_effect_direction}
\title{Title}
\usage{
tri_swap_effect_direction(dat, types = NULL)
}
\arguments{
\item{dat}{Data in long format}

\item{types}{Vector of study types for which the direction should be reversed}
}
\description{
Title
}
