% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_ind}
\alias{dat_ind}
\title{DATASET_TITLE}
\format{
A data frame with 20 rows and 14 variables:
\describe{
\item{\code{result_id}}{character COLUMN_DESCRIPTION}
\item{\code{study}}{character COLUMN_DESCRIPTION}
\item{\code{type}}{character COLUMN_DESCRIPTION}
\item{\code{yi}}{double COLUMN_DESCRIPTION}
\item{\code{vi}}{double COLUMN_DESCRIPTION}
\item{\code{d1j}}{character COLUMN_DESCRIPTION}
\item{\code{d2j}}{character COLUMN_DESCRIPTION}
\item{\code{d3j}}{character COLUMN_DESCRIPTION}
\item{\code{d1t}}{character COLUMN_DESCRIPTION}
\item{\code{d2t}}{character COLUMN_DESCRIPTION}
\item{\code{d3t}}{character COLUMN_DESCRIPTION}
\item{\code{d1d}}{character COLUMN_DESCRIPTION}
\item{\code{d2d}}{character COLUMN_DESCRIPTION}
\item{\code{d3d}}{character COLUMN_DESCRIPTION}
}
}
\usage{
dat_ind
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
