% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_ind_values}
\alias{dat_ind_values}
\title{DATASET_TITLE}
\format{
A data frame with 3 rows and 6 variables:
\describe{
\item{\code{domain}}{character COLUMN_DESCRIPTION}
\item{\code{j}}{character COLUMN_DESCRIPTION}
\item{\code{ind_m_add}}{double COLUMN_DESCRIPTION}
\item{\code{ind_v_add}}{double COLUMN_DESCRIPTION}
\item{\code{ind_m_prop}}{double COLUMN_DESCRIPTION}
\item{\code{ind_v_prop}}{double COLUMN_DESCRIPTION}
}
}
\usage{
dat_ind_values
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
