% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_direction.R
\name{tri_absolute_direction_quick}
\alias{tri_absolute_direction_quick}
\title{Quick wrapper to apply absolute direction transformation}
\usage{
tri_absolute_direction_quick(dat, invert = FALSE)
}
\arguments{
\item{dat}{Dataset in wide format ## check this is the correct parameter}

\item{invert}{If TRUE, reverse transformation (absolute → qualitative)}
}
\description{
Converts data between wide and long format while applying or reversing
the absolute direction transformation.
}
