% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_values.R
\name{tri_append_bias}
\alias{tri_append_bias}
\title{Append bias priors to bias dataset}
\usage{
tri_append_bias(dat, values, common = TRUE)
}
\arguments{
\item{dat}{A bias dataset in long format (must include j, d, t, domain, result_id)}

\item{values}{A data frame of prior distributions (e.g., from dat_bias_values)}

\item{common}{Should a single set of prior distributions be used across all domains (default is TRUE)? Set to FALSE to allow domain-specific priors.}
}
\value{
Dataset with numeric bias prior values appended
}
\description{
Joins the appropriate prior parameters to a long-format bias dataset, based on
the risk-of-bias judgment (j), direction (d), and type (t).
}
