% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_values.R
\name{tri_append_indirect}
\alias{tri_append_indirect}
\title{Append indirectness priors to dataset}
\usage{
tri_append_indirect(dat, values, common = TRUE)
}
\arguments{
\item{dat}{A long-format indirectness dataset}

\item{values}{A dataframe of indirectness priors (e.g., dat_ind_values)}

\item{common}{Should a single set of distributions be used across all domains (default is TRUE)?}
}
\value{
Dataset with indirectness prior values appended
}
\description{
Similar to tri_append_bias(), but operates on indirectness judgments and priors
}
