% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{tri_calculate_adjusted_estimates}
\alias{tri_calculate_adjusted_estimates}
\title{Title}
\usage{
tri_calculate_adjusted_estimates(dat)
}
\arguments{
\item{dat}{Data in long format}
}
\description{
Title
}
