% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{tri_calculate_adjusted_estimates}
\alias{tri_calculate_adjusted_estimates}
\title{Calculate Bias-Adjusted Estimates}
\usage{
tri_calculate_adjusted_estimates(dat)
}
\arguments{
\item{dat}{A dataset prepared using \code{tri_prep_data()}, containing bias parameters}
}
\value{
A tibble with added \code{yi_adj} and \code{vi_adj} columns
}
\description{
Computes adjusted effect estimates (\code{yi_adj}) and variances (\code{vi_adj})
incorporating additive and proportional bias adjustments.
}
