% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_direction.R
\name{tri_swap_effect_direction}
\alias{tri_swap_effect_direction}
\title{Standardise Effect Direction and Adjust Bias Directions}
\usage{
tri_swap_effect_direction(dat, types = NULL)
}
\arguments{
\item{dat}{A long-format triangulation dataset}

\item{types}{A character vector of \code{type} values to flip (e.g. c("OR", "RR"))}
}
\value{
A modified dataset with adjusted \code{yi} and \code{d} values
}
\description{
Flips effect estimates (\code{yi}) and proportional bias directions (\code{d})
for specified study types (e.g., odds ratios) to ensure consistency
across studies using different metrics.
}
