% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulate-package.R
\docType{package}
\name{triangulate-package}
\alias{triangulate}
\alias{triangulate-package}
\title{triangulate: Useful functions for performing evidence triangulation}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mcguinlu.github.io/triangulate}
  \item \url{https://github.com/mcguinlu/triangulate}
  \item \url{https://mrcieu.r-universe.dev/triangulate}
  \item Report bugs at \url{https://github.com/mcguinlu/triangulate/issues}
}

}
\author{
\strong{Maintainer}: Luke McGuinness \email{luke.mcguinness@bristol.ac.uk} (\href{https://orcid.org/0000-0001-8730-9761}{ORCID})

Authors:
\itemize{
  \item Tassia Jones \email{nv22679@bristol.ac.uk}
  \item Chin Yang Shapland \email{chinyang.shapland@bristol.ac.uk}
}

}
\keyword{internal}
