% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulate.R
\name{triangulate}
\alias{triangulate}
\title{Connect 5 points to form 3 triangles}
\usage{
triangulate(points)
}
\arguments{
\item{points}{a 5x2 matrix containing 5 points}
}
\value{
A list containing points (vertices) and edges of triangles
}
\description{
This function that creates three 3 triangles
by connecting any five (5) points in a 2D plane
following the cartesian coordinates system
}
\examples{
points <- matrix(c(0,0,1,1,2,0,3,1,4,0), nrow=5, byrow=T)
triangulate(points)
}
